TrafficML-Simulation/requirements.txt
Flask
Flask-MQTT
paho-mqtt
numpy
pandas
scikit-learn
traci
joblib
